Canvas generation update
Updated the way each 'canvas' works - they are now actually a collection of 4 images. This has allowed me to ensure that the images do not appear in the wrong order (the character on top of the speech box, for example). It is also more efficient as I only need to update the correct layer rather than regenerating all of them.