Improved mationality filtering
I went through the list of all TGAA characters and added all the nationalities to the list. Despite not having most of these characters available yet, this forced me to consider what to do if I did. I've now got this to a working state, although I plan on improving it in the next commit.