Nationality amalgamation
The previous nationality filtering was a bit cumbersome and unnecessarily complicated. While the vast majority of characters are either British or Japanese, there are a handful of other nationalities with only one or two characters each. Rather than having checkboxes for all of these, I have added an 'other' category.